moved additional pages out of admin
