+++ src/domain/journey/opportunity/pages/OpportunityCommunications/OpportunityCommunicationsPage.tsx
@@ -1,7 +1,5 @@
 import { PropsWithChildren } from 'react';
-import CommunityUpdatesPage, {
-  CommunityUpdatesPageProps,
-} from '@/domain/platform/admin/community/CommunityUpdatesPage';
+import CommunityUpdatesPage, { CommunityUpdatesPageProps } from '@/domain/space/pages/CommunityUpdatesPage';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
