+++ src/domain/journey/space/pages/SpaceAccount/SpaceAccountPage.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Skeleton } from '@mui/material';
 import useUrlResolver from '@/main/routing/urlResolver/useUrlResolver';
-import SpaceSettingsLayout from '@/domain/platform/admin/space/SpaceSettingsLayout';
+import SpaceSettingsLayout from '@/domain/space/admin/SpaceSettingsLayout';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import SpaceAccountView from './SpaceAccountView';
