+++ src/domain/journey/space/pages/SpaceSettings/SpaceSettingsPage.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import SpaceSettingsView from './SpaceSettingsView';
-import SpaceSettingsLayout from '@/domain/platform/admin/space/SpaceSettingsLayout';
+import SpaceSettingsLayout from '@/domain/space/admin/SpaceSettingsLayout';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import useUrlResolver from '@/main/routing/urlResolver/useUrlResolver';
