+++ src/domain/journey/subspace/pages/SubspaceCommunications/SubspaceCommunicationsPage.tsx
@@ -1,9 +1,7 @@
 import React, { FC } from 'react';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
-import CommunityUpdatesPage, {
-  CommunityUpdatesPageProps,
-} from '@/domain/platform/admin/community/CommunityUpdatesPage';
+import CommunityUpdatesPage, { CommunityUpdatesPageProps } from '@/domain/space/pages/CommunityUpdatesPage';
 import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
 
 interface SubspaceCommunicationsPageProps extends SettingsPageProps, CommunityUpdatesPageProps {}
