+++ src/domain/space/admin/SpaceTemplatesAdminPage.tsx
@@ -4,8 +4,8 @@ import useUrlResolver from '@/main/routing/urlResolver/useUrlResolver';
 import Loading from '@/core/ui/loading/Loading';
 import { buildSettingsUrl } from '@/main/routing/urlBuilders';
 import TemplatesAdmin from '@/domain/templates/components/TemplatesAdmin/TemplatesAdmin';
-import { SettingsSection } from '../layout/EntitySettingsLayout/SettingsSection';
-import { SettingsPageProps } from '../layout/EntitySettingsLayout/types';
+import { SettingsSection } from '../../platform/admin/layout/EntitySettingsLayout/SettingsSection';
+import { SettingsPageProps } from '../../platform/admin/layout/EntitySettingsLayout/types';
 import SpaceSettingsLayout from './SpaceSettingsLayout';
 import { TemplateType } from '@/core/apollo/generated/graphql-schema';
 import { useSpace } from '@/domain/space/SpaceContext/useSpace';
