+++ src/domain/space/admin/SpaceTemplatesAdminRoutes.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes, useResolvedPath } from 'react-router-dom';
 import SpaceTemplatesAdminPage from './SpaceTemplatesAdminPage';
-import { SettingsPageProps } from '../layout/EntitySettingsLayout/types';
+import { SettingsPageProps } from '../../platform/admin/layout/EntitySettingsLayout/types';
 import { nameOfUrl } from '@/main/routing/urlParams';
 
 interface SpaceTemplatesAdminRoutesProps extends SettingsPageProps {
