+++ src/domain/space/admin/storage/SpaceStorageAdminQueries.graphql
