+++ src/domain/space/routing/toReview2/SpaceSettingsRoute.tsx
@@ -6,8 +6,8 @@ import { Error404 } from '@/core/pages/Errors/Error404';
 import SpaceCommunicationsPage from '@/domain/journey/space/pages/SpaceCommunication/SpaceCommunicationsPage';
 import SpaceSettingsAboutPage from '@/domain/journey/space/pages/SpaceAboutSettings/SpaceAboutPage';
 import SpaceSettingsPage from '@/domain/journey/space/pages/SpaceSettings/SpaceSettingsPage';
-import SpaceTemplatesAdminRoutes from '@/domain/platform/admin/space/SpaceTemplatesAdminRoutes';
-import SpaceStorageAdminPage from '@/domain/platform/admin/space/storage/SpaceStorageAdminPage';
+import SpaceTemplatesAdminRoutes from '@/domain/space/admin/SpaceTemplatesAdminRoutes';
+import SpaceStorageAdminPage from '@/domain/space/admin/storage/SpaceStorageAdminPage';
 import { StorageConfigContextProvider } from '@/domain/storage/StorageBucket/StorageConfigContext';
 import AdminSpaceCommunityPage from '@/domain/journey/space/pages/AdminSpaceCommunityPage';
 import SpaceAccountPage from '@/domain/journey/space/pages/SpaceAccount/SpaceAccountPage';
